+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Bump flipper to v0.137.0
 - Bump GH Actions `upload-artifact` to v3
 - Bump logback to v1.2.11
+- Move wildcard formatting to search query
 - [In Progress: 17 Feb 2022] Add `CriticalAppUpdateDialog`
 
 ### Changes